+++ packages/xo-web/src/common/select-objects.js
@@ -86,6 +86,9 @@ const options = props => ({
   defaultValue: props.multi ? [] : undefined,
 })
 
+const getObjectsById = objects =>
+  keyBy(isArray(objects) ? objects : flatten(toArray(objects)), 'id')
+
 // ===================================================================
 
 /*
@@ -126,15 +129,50 @@ class GenericSelect extends React.Component {
     ]).isRequired,
   }
 
-  _getObjectsById = createSelector(
-    () => this.props.xoObjects,
-    objects =>
-      keyBy(isArray(objects) ? objects : flatten(toArray(objects)), 'id')
+  _getSelectedIds = createSelector(
+    () => this.props.value,
+    createCollectionWrapper(getIds)
   )
+
+  _getObjects = createSelector(
+    () => this.props.xoContainers !== undefined,
+    () => this.props.xoObjects,
+    this._getSelectedIds,
+    (withContainers, objects, ids) => {
+      const objectsById = getObjectsById(objects)
+      const missingObjects = []
+      const addIfMissing = id => {
+        if (id != null && !(id in objectsById)) {
+          missingObjects.push({
+            id,
+            label: id,
+            value: id,
+            removed: true,
+          })
+        }
+      }
+      if (isArray(ids)) {
+        ids.forEach(addIfMissing)
+      } else {
+        addIfMissing(ids)
+      }
+
+      return isEmpty(missingObjects)
+        ? objects
+        : withContainers
+          ? {
+              ...objects,
+              missingObjects,
+            }
+          : [...objects, ...missingObjects]
+    }
+  )
+
+  _getObjectsById = createSelector(this._getObjects, getObjectsById)
 
   _getOptions = createSelector(
     () => this.props.xoContainers,
-    () => this.props.xoObjects,
+    this._getObjects,
     (containers, objects) => {
       // createCollectionWrapper with a depth?
       const { name } = this.constructor
@@ -166,38 +204,24 @@ class GenericSelect extends React.Component {
             options.push(getOption(object, container))
           })
         })
-      }
 
-      const objectsById = this._getObjectsById()
-      const addIfMissing = val => {
-        if (val != null && !(val in objectsById)) {
-          options.push({
-            disabled: true,
-            id: val,
-            label: val,
-            value: val,
-            xoItem: {
-              id: val,
-              removed: true,
-            },
+        // missing objects have "missingObjects" as container
+        const { missingObjects } = objects
+        if (missingObjects !== undefined) {
+          missingObjects.forEach(object => {
+            options.push(getOption(object))
           })
         }
       }
 
-      const values = this._getSelectedIds()
-      if (isArray(values)) {
-        forEach(values, addIfMissing)
-      } else {
-        addIfMissing(values)
-      }
-
+      options.map(option => {
+        if (option.xoItem.removed) {
+          option.disabled = true
+        }
+        return option
+      })
       return options
     }
-  )
-
-  _getSelectedIds = createSelector(
-    () => this.props.value,
-    createCollectionWrapper(getIds)
   )
 
   _getSelectedObjects = (() => {